[Reporting] close reporting after archiving it and move useCases in feature
