+++ frontend/src/domain/use_cases/missions/deleteTab.ts
@@ -1,10 +1,10 @@
+import { getIdTyped } from '@features/missions/utils'
 import { generatePath } from 'react-router'
 
 import { attachReportingToMissionSliceActions } from '../../../features/missions/MissionForm/AttachReporting/slice'
 import { missionFormsActions, type MissionInStateType } from '../../../features/missions/MissionForm/slice'
 import { missionActions } from '../../../features/missions/slice'
 import { sideWindowActions } from '../../../features/SideWindow/slice'
-import { getIdTyped } from '../../../utils/getIdTyped'
 import { getMissionPageRoute } from '../../../utils/routes'
 import { sideWindowPaths } from '../../entities/sideWindow'
