+++ frontend/src/domain/use_cases/missions/switchTab.ts
@@ -1,8 +1,9 @@
+import { getIdTyped } from '@features/missions/utils'
+
 import { attachReportingToMissionSliceActions } from '../../../features/missions/MissionForm/AttachReporting/slice'
 import { missionFormsActions } from '../../../features/missions/MissionForm/slice'
 import { missionActions } from '../../../features/missions/slice'
 import { sideWindowActions } from '../../../features/SideWindow/slice'
-import { getIdTyped } from '../../../utils/getIdTyped'
 import { getMissionPageRoute } from '../../../utils/routes'
 
 import type { HomeAppThunk } from '@store/index'
