+++ frontend/src/features/Reportings/components/ReportingForm/AttachMission/AttachedMissionCard.tsx
@@ -1,9 +1,8 @@
 import { CompletionStatusTag } from '@features/missions/components/CompletionStatusTag'
 import { MissionStatusTag } from '@features/missions/components/MissionStatusTag'
-import { getMissionCompletionStatus } from '@features/missions/utils'
+import { getMissionCompletionStatus, humanizeMissionTypes } from '@features/missions/utils'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { customDayjs as dayjs, pluralize } from '@mtes-mct/monitor-ui'
-import { humanizeMissionTypes } from '@utils/humanizeMissionTypes'
 import { getMissionStatus, type Mission } from 'domain/entities/missions'
 import { editMissionInLocalStore } from 'domain/use_cases/missions/editMissionInLocalStore'
 import styled from 'styled-components'
