+++ frontend/src/features/Reportings/components/ReportingLayer/MissionToAttach/index.tsx
@@ -7,11 +7,11 @@ import { useEffect, useMemo, useRef, type MutableRefObject } from 'react'
 import { useGetMissionsQuery } from '../../../../../api/missionsAPI'
 import { Layers } from '../../../../../domain/entities/layers/constants'
 import { MissionStatusEnum } from '../../../../../domain/entities/missions'
-import { attachMission } from '../../../../../domain/use_cases/reporting/attachMission'
 import { useAppDispatch } from '../../../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../../../hooks/useAppSelector'
 import { getMissionZoneFeature } from '../../../../missions/Layers/missionGeometryHelpers'
 import { missionStyleFn } from '../../../../missions/Layers/missions.style'
+import { attachMission } from '../../../useCases/attachMission'
 
 import type { VectorLayerWithName } from '../../../../../domain/types/layer'
 import type { BaseMapChildrenProps } from '../../../../map/BaseMap'
