+++ frontend/src/features/Reportings/components/ReportingsList/Cells/ButtonsRowGroup.tsx
@@ -1,13 +1,13 @@
 import { ArchiveModal } from '@features/commonComponents/Modals/Archive'
 import { DeleteModal } from '@features/commonComponents/Modals/Delete'
+import { archiveReportingFromTable } from '@features/Reportings/useCases/archiveReporting'
+import { deleteReporting } from '@features/Reportings/useCases/deleteReporting'
+import { duplicateReporting } from '@features/Reportings/useCases/duplicateReporting'
+import { editReportingInLocalStore } from '@features/Reportings/useCases/editReportingInLocalStore'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { useAppSelector } from '@hooks/useAppSelector'
 import { Accent, Dropdown, Icon, IconButton } from '@mtes-mct/monitor-ui'
 import { ReportingContext } from 'domain/shared_slices/Global'
-import { archiveReportingFromTable } from 'domain/use_cases/reporting/archiveReporting'
-import { deleteReporting } from 'domain/use_cases/reporting/deleteReporting'
-import { duplicateReporting } from 'domain/use_cases/reporting/duplicateReporting'
-import { editReportingInLocalStore } from 'domain/use_cases/reporting/editReportingInLocalStore'
 import { useState } from 'react'
 import styled from 'styled-components'
