+++ frontend/src/features/Reportings/hooks/useGetFilteredReportingsQuery.ts
@@ -5,10 +5,10 @@ import { useMemo } from 'react'
 
 import { useGetReportingsQuery } from '../../../api/reportingsAPI'
 import { TWO_MINUTES } from '../../../constants'
-import { sourceFilterFunction } from '../../../domain/use_cases/reporting/filters/sourceFilterFunction'
-import { subThemesFilterFunction } from '../../../domain/use_cases/reporting/filters/subThemesFilterFunction'
-import { themeFilterFunction } from '../../../domain/use_cases/reporting/filters/themeFilterFunction'
 import { useAppSelector } from '../../../hooks/useAppSelector'
+import { sourceFilterFunction } from '../useCases/filters/sourceFilterFunction'
+import { subThemesFilterFunction } from '../useCases/filters/subThemesFilterFunction'
+import { themeFilterFunction } from '../useCases/filters/themeFilterFunction'
 
 export const useGetFilteredReportingsQuery = (skip = false) => {
   const {
