+++ frontend/src/features/Reportings/useCases/addReporting.ts
@@ -1,11 +1,11 @@
 import { attachMissionToReportingSliceActions } from '@features/Reportings/components/ReportingForm/AttachMission/slice'
 import { reportingActions } from '@features/Reportings/slice'
+import { ReportingContext, setReportingFormVisibility, VisibilityState } from 'domain/shared_slices/Global'
 
-import { mainWindowActions } from '../../../features/MainWindow/slice'
-import { getReportingInitialValues, createIdForNewReporting } from '../../../features/Reportings/utils'
-import { setReportingFormVisibility, ReportingContext, VisibilityState } from '../../shared_slices/Global'
+import { mainWindowActions } from '../../MainWindow/slice'
+import { getReportingInitialValues, createIdForNewReporting } from '../utils'
 
-import type { Reporting } from '../../entities/reporting'
+import type { Reporting } from '../../../domain/entities/reporting'
 import type { HomeAppThunk } from '@store/index'
 
 export const addReporting =
