+++ frontend/src/features/Reportings/useCases/deleteReportings.ts
@@ -1,9 +1,9 @@
 import { attachMissionToReportingSliceActions } from '@features/Reportings/components/ReportingForm/AttachMission/slice'
 import { reportingActions } from '@features/Reportings/slice'
+import { ReportingContext, setReportingFormVisibility, setToast, VisibilityState } from 'domain/shared_slices/Global'
 
 import { reportingsAPI } from '../../../api/reportingsAPI'
-import { mainWindowActions } from '../../../features/MainWindow/slice'
-import { ReportingContext, setReportingFormVisibility, setToast, VisibilityState } from '../../shared_slices/Global'
+import { mainWindowActions } from '../../MainWindow/slice'
 
 export const deleteReportings = (ids: number[], resetSelectionFn: () => void) => async (dispatch, getState) => {
   const { activeReportingId, context, selectedReportingIdOnMap } = getState().reporting
