+++ frontend/src/features/Reportings/useCases/duplicateReporting.ts
@@ -1,9 +1,9 @@
 import { attachMissionToReportingSliceActions } from '@features/Reportings/components/ReportingForm/AttachMission/slice'
 import { reportingActions } from '@features/Reportings/slice'
+import { setReportingFormVisibility, setToast, ReportingContext, VisibilityState } from 'domain/shared_slices/Global'
 
 import { reportingsAPI } from '../../../api/reportingsAPI'
-import { getReportingInitialValues, createIdForNewReporting } from '../../../features/Reportings/utils'
-import { setReportingFormVisibility, setToast, ReportingContext, VisibilityState } from '../../shared_slices/Global'
+import { getReportingInitialValues, createIdForNewReporting } from '../utils'
 
 export const duplicateReporting = (reportingId: number) => async (dispatch, getState) => {
   const { reportings } = getState().reporting
