+++ frontend/src/features/Reportings/useCases/filters/subThemesFilterFunction.ts
