+++ frontend/src/features/Reportings/useCases/reduceReportingFormOnMap.ts
@@ -1,5 +1,6 @@
-import { mainWindowActions } from '../../../features/MainWindow/slice'
-import { setReportingFormVisibility, ReportingContext, VisibilityState } from '../../shared_slices/Global'
+import { setReportingFormVisibility, ReportingContext, VisibilityState } from 'domain/shared_slices/Global'
+
+import { mainWindowActions } from '../../MainWindow/slice'
 
 import type { HomeAppThunk } from '../../../store'
