+++ frontend/src/features/Semaphore/components/Overlay/SemaphoreCard.tsx
@@ -1,4 +1,5 @@
 import { reportingsFiltersActions, ReportingsFiltersEnum } from '@features/Reportings/Filters/slice'
+import { addReporting } from '@features/Reportings/useCases/addReporting'
 import { sideWindowActions } from '@features/SideWindow/slice'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { useAppSelector } from '@hooks/useAppSelector'
@@ -10,7 +11,6 @@ import { sideWindowPaths } from 'domain/entities/sideWindow'
 import { removeOverlayStroke, ReportingContext } from 'domain/shared_slices/Global'
 import { resetSelectedSemaphore } from 'domain/shared_slices/SemaphoresSlice'
 import { closeAllOverlays } from 'domain/use_cases/map/closeAllOverlays'
-import { addReporting } from 'domain/use_cases/reporting/addReporting'
 import { useCallback, useState } from 'react'
 import { Tooltip, Whisper } from 'rsuite'
 import styled from 'styled-components'
