+++ frontend/src/features/missions/MissionForm/index.tsx
@@ -7,9 +7,9 @@ import styled from 'styled-components'
 import { MissionForm } from './MissionForm'
 import { MissionSchema } from './Schemas'
 import { useAppSelector } from '../../../hooks/useAppSelector'
-import { isNewMission } from '../../../utils/isNewMission'
 import { missionFactory } from '../Missions.helpers'
 import { getActiveMission } from './slice'
+import { isMissionNew } from '../utils'
 
 import type { Mission as MissionType, NewMission } from '../../../domain/entities/missions'
 
@@ -22,7 +22,7 @@ export function MissionFormWrapper() {
 
   const activeAction = selectedMission?.activeAction
 
-  const missionIsNewMission = useMemo(() => isNewMission(activeMissionId), [activeMissionId])
+  const missionIsNewMission = useMemo(() => isMissionNew(activeMissionId), [activeMissionId])
 
   const missionValues: Partial<MissionType> = useMemo(() => {
     if (selectedMission?.missionForm) {
