+++ frontend/src/features/missions/MissionsButton/index.tsx
@@ -1,5 +1,6 @@
 import { DialogButton, StyledMapMenuDialogContainer } from '@components/style'
 import { MenuWithCloseButton } from '@features/commonStyles/map/MenuWithCloseButton'
+import { reduceReportingFormOnMap } from '@features/Reportings/useCases/reduceReportingFormOnMap'
 import { SideWindowStatus, sideWindowActions } from '@features/SideWindow/slice'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { useAppSelector } from '@hooks/useAppSelector'
@@ -8,7 +9,6 @@ import { isMissionOrMissionsPage } from '@utils/routes'
 import { sideWindowPaths } from 'domain/entities/sideWindow'
 import { globalActions, setDisplayedItems } from 'domain/shared_slices/Global'
 import { addMission } from 'domain/use_cases/missions/addMission'
-import { reduceReportingFormOnMap } from 'domain/use_cases/reporting/reduceReportingFormOnMap'
 import { useMemo } from 'react'
 
 import { MissionFilterContext, MissionFilters } from '../components/Filters'
