+++ frontend/src/features/missions/Overlays/MissionCard.tsx
@@ -2,11 +2,10 @@ import { Bold } from '@components/style'
 import { CompletionStatusTag } from '@features/missions/components/CompletionStatusTag'
 import { MissionStatusTag } from '@features/missions/components/MissionStatusTag'
 import { missionActions } from '@features/missions/slice'
-import { getMissionCompletionStatus } from '@features/missions/utils'
+import { getMissionCompletionStatus, humanizeMissionTypes } from '@features/missions/utils'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { useAppSelector } from '@hooks/useAppSelector'
 import { Accent, Button, Icon, IconButton, Size, customDayjs as dayjs, pluralize } from '@mtes-mct/monitor-ui'
-import { humanizeMissionTypes } from '@utils/humanizeMissionTypes'
 import { removeOverlayStroke } from 'domain/shared_slices/Global'
 import { closeAllOverlays } from 'domain/use_cases/map/closeAllOverlays'
 import { editMissionInLocalStore } from 'domain/use_cases/missions/editMissionInLocalStore'
