--- frontend/src/utils/getIdTyped.ts
@@ -1,7 +1,0 @@
-export function getIdTyped(id: string | undefined) {
-  if (!id) {
-    return undefined
-  }
-
-  return id.includes('new-') ? id : Number(id)
-}
