--- frontend/src/utils/isNewMission.ts
@@ -1,3 +1,0 @@
-export function isNewMission(id: string | number | undefined) {
-  return id?.toString().includes('new') ?? false
-}
